requirements.txt
alembic==0.9.5
aniso8601==1.2.1
appnope==0.1.0
asn1crypto==0.24.0
attrs==17.4.0
bcrypt==3.1.4
bleach==2.1.2
boto3==1.6.4
botocore==1.9.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
cryptography==2.2.1
decorator==4.2.1
docutils==0.14
entrypoints==0.2.3
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JWT-Extended==3.10.0
Flask-Migrate==2.1.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-Testing==0.7.1
future==0.16.0
futures==3.1.1
html5lib==1.0.1
idna==2.6
ipaddress==1.0.14
ipykernel==4.8.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Mako==1.0.7
MarkupSafe==1.0
mistune==0.8.3
mysql-connector-python==8.0.6
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
paramiko==2.4.1
parso==0.1.1
passlib==1.7.1
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
py4j==0.10.6
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyHive==0.5.1
PyJWT==1.4.2
pykerberos==1.2.1
PyNaCl==1.2.1
pyspark==2.3.0
pyspin==1.1.1
pytest==3.4.2
pytest-arraydiff==0.2
pytest-doctestplus==0.1.2
pytest-openfiles==0.2.0
pytest-remotedata==0.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.13.0
requests-kerberos==0.10.0
s3transfer==0.1.13
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.10.0
slackclient==1.2.1
SQLAlchemy==1.1.13
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typing==3.5.3.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.47.0
Werkzeug==0.14.1
widgetsnbextension==3.1.3
xlrd==1.1.0